agentfabric: Add new lint rules
