--- dialect/agentfabric/src/compiler/index.ts
@@ -1,22 +1,0 @@
-export { compile } from './compile.js';
-export type { CompileResult, CompileOptions } from './compile.js';
-export type { AgentGraph } from './agent-graph.js';
-export type {
-  UnifiedAgentSpecification,
-  AgentNode,
-  ActionNode,
-  RouterNode,
-  HandoffAction,
-  ActionCallableReference,
-  ActionDefinition,
-  LLMRef,
-  StateVariable,
-  GraphConfig,
-  Node,
-} from './unified-agent-specification.js';
-export { ObjectTypes } from './unified-agent-specification.js';
-export type { LLMProvider, InvokableClient } from './service-types.js';
-export type {
-  CompilerDiagnostic,
-  DiagnosticSeverity,
-} from './compiler-context.js';
