+++ dialect/agentfabric/src/lint/passes/rules/action-binding-rules.ts
@@ -15,11 +15,11 @@ import {
   decomposeAtMemberExpression,
   isNamedMap,
 } from '@agentscript/language';
-import { normalizeId } from '../../../compiler/utils.js';
 import {
+  normalizeId,
   IMPLICIT_WITH_PARAMS,
   listActionDefInputNames,
-} from '../../../compiler/build-nodes.js';
+} from '../../utils.js';
 import { attachError, extractStringValue, type AstLike } from './shared.js';
 
 function getActionDefName(
