+++ dialect/agentfabric/src/tests/llm-value-completions.test.ts
@@ -2,7 +2,7 @@
  * Regression tests: value-position completions for enum-typed fields inside
  * an LLM entry should include the enum members.
  *
- * Bug (W-22415806): when the cursor is at value position (after `key: `) for
+ * Bug: when the cursor is at value position (after `key: `) for
  * an enum-typed field, the LSP returns no completions for:
  *   - `kind:` → expected to suggest the discriminator enum members
  *     (OpenAI, Gemini)
